Improve icon input on comments.
If no current text selection, move the caret where data must be entered
next. Else insert the selected text where it should be and retain selection.

Multiline selections are checked for use of alternate delimiters and/or
line separators.

On MSIE, latch text selection on mousedown event, because it is cleared
before click event occurs.