Support for issue comment notification by e-mail and subscription.
Users can be subscribed to an issue by creating the issue, commenting,
being mentioned or owning the repository. At any time they can
explicitly subscribe/unsubscribe via a button on the issue page.

Notifications can be sent either by SMTP or the external sendmail
program. Note that esmtp's sendmail is not suitable for that use.

PR handling is not (yet) ready for notifications. Likewise, it is not
possible to comment an issue by replying to a notification mail.