PR diff: use merge base as start point instead of base head
Currently, a PR's file count and diff wrongly reflects changes made to
the base branch in the meantime.
This commit gets the merge base point between the base and target
branches and uses it as the starting point for PR's commits. The merge
base is normally invariant, so does the file count and diff shown.