js: implement form default submit.
Upon typing the enter key in a form, browsers may submit the latter as if
the first submit button has been clicked: this may not be what is
expected by the user and this commit implements a way to use another
default submit button.

By including the non-standard attribute defaultSubmit="selector" in a
form tag, an additional non-visible button captures these default submit
events and redirects them to the selected button. Forms are
automatically preprocessed by Gitprep's javascript to service this new
attribute.