Reduce git push time by speeding up git hooks start
Loading the Gitprep application in a program delays considerably its
start time: measures performed locally show that, for a simple git push
taking an overall elapsed time of 2.5 seconds, 1.5 seconds are spent
loading the application. The application is loaded 3 times (gitprep
shell and both hooks), so we can consider each loading takes ~0.5
seconds for this particular use case.

This commit introduces a kind of RPC from the hooks, in order to avoid
them loading the Gitprep application: as the Gitprep shell is alive
when hooks are executed, it implements the RPC server and executes
requests from the hooks on its own loaded application.

The communication channel is an orphan socket open on file descriptor 99,
allowing only programs scheduled by the git shell to access it, thus no
foreign program can inherit or open it.

Since the channel is a single-preopened socket, it cannot serve requests
from more than one client at a time: this is not a problem in our case,
as hooks are not executed concurrently.

The Gitprep shell do not wait passively on the git shell anymore, but
serves RPC requests while monitoring the git shell output to detect its
exit. No additional thread is created.

RPC implementation is not a standard, but designed specifically for this
case: requests and replies are transmitted as JSON messages terminated
by a newline. Requests hold a Perl-like expression based on the app
reference and a hash defining the client scalar/reference variables
used by the expression. Replies contain either the expression value or
an error message.

A Gitprep::RPC module is introduced for that purpose. It is not loaded
by the application itself.

As such, this commit reduces the above measurement case from about 2.5
to 1.5 seconds.