Fix smart http pushes
Introduction of git hooks broke http pushes, because the formers were
not executed with the necessary environment variables.

This commit moves hooks execution preparing from the gitprep shell to
the manager api and calls it from both the gitprep shell and smart http
response code. It makes sure a "session" user id is defined for the
hooks.