Work on branch page
- Add branch table header line
- Rework line layout and controls
- Feature branch creation
- Feature branch renaming

Take rulesets into considerations when creating/deleting/renaming a
branch. A branch cannot be renamrd if it's locked for deletion and/or
its new name is not allowed for creation.

Adds a branch name check in custom validation.

Also do not allow creating a user project with name ending in ".wiki".